+++ android/app/src/main/java/app/williamtemple/lotto/MainActivity.java
@@ -1,0 +1,5 @@
+package app.williamtemple.lotto;
+
+import com.getcapacitor.BridgeActivity;
+
+public class MainActivity extends BridgeActivity {}